requirements.txt
requests==2.31.0
python-dotenv==1.0.0
playwright==1.42.0
beautifulsoup4==4.12.3
pymongo==4.6.2
google-generativeai==0.8.3 